No Issue - Sync up Windows Installer with Pale Moon.
Remove unused DEV_EDITION conditionals.
Stop hardcoding installer app name values.

This also fixes the installer's broken detection of a running Pale Moon
instance. Supposedly, the installer will prompt the user to close the
application if it finds a matching window class. Since the window class
of the browser uses the remoting name, not the display name
(current value), the installer will note that a restart will be required
instead of showing the prompt.